// base/free5gc/webconsole/frontend/src/api/api.ts
export interface ChargingData {
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'snssai'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'dnn': string;
    /**
     * 
     * @type {number}
     * @memberof ChargingData
     */
    'qosRef'?: number;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'filter': string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'chargingMethod'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'quota'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'unitCost'?: string;
    /**
     * 
     * @type {string}
     * @memberof ChargingData
     */
    'ueId'?: string;
}
/**
 * 
 * @export
 * @interface DefaultSingleNssais
 */
export interface DefaultSingleNssais {
    /**
     * 
     * @type {Array<Nssai>}
     * @memberof DefaultSingleNssais
     */
    'defaultSingleNssais': Array<Nssai>;
    /**
     * 
     * @type {Array<Nssai>}
     * @memberof DefaultSingleNssais
     */
    'singleNssais'?: Array<Nssai>;
}
/**
 * 
 * @export
 * @interface Dnn
 */
export interface Dnn {
    /**
     * 
     * @type {string}
     * @memberof Dnn
     */
    'dnn': string;
}
// ...
export interface FlowChargingRecord {
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'Supi': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'Snssai': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'Dnn': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'Filter': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'QuotaLeft': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'Usage': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'TotalVol': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'UlVol': string;
    /**
     * 
     * @type {string}
     * @memberof FlowChargingRecord
     */
    'DlVol': string;
}